Edit base_model class to use some storage methods
If kwargs are not passed public attr are set and storage.new is call.
This is done to create a new obj into our storage __objects attribute

From save method we call our storage.save method.
This writes the new instance to our JSON file that contains all objs

A change was made to our to_dict method as it was editing  __dict__.
This caused an issue when calling storage.new as it uses to_dict method.
More info on comment added to method